requirements.txt
dnspython==2.1.0
Flask==1.1.2
Flask-Cors==3.0.10
Flask-PyMongo==2.3.0
Flask-Session==0.3.2
Flask-Mail==0.9.1
blinker==1.4
future==0.18.2
json5==0.9.5
jsonschema==3.2.0
path==13.1.0
pymongo==3.11.3
argon2-cffi==20.1.0